feat: optimize scripts and docs (#14)
* repo-sync-2023-07-07T15:33:54+0800

* revert change

* fix comma

+++ scripts/deploy/init_kusciaapi_cert.sh
@@ -58,8 +58,10 @@ openssl x509  -req -in ${CLIENT}.csr \
     -out ${CLIENT}.crt
 
 #generate token file
-openssl rand -base64 8 |xargs echo -n > /tmp/token
-sha256sum /tmp/token | cut -d' ' -f1 |xargs echo -n > token
-rm -rf /tmp/token
+if [ ! -e token ]; then # token not exists
+  openssl rand -base64 8 |xargs echo -n > /tmp/token
+  sha256sum /tmp/token | cut -d' ' -f1 |xargs echo -n > token
+  rm -rf /tmp/token
+fi
 
 popd || exit
